// frontend/LawyerSite.Web/src/app/services/client/web-api-client.ts
export interface ServiceDto {
    id?: number;
    title?: string;
    emoji?: string | undefined;
    description?: string | undefined;
    price?: number;
    isActive?: boolean;
    durationMinutes?: number;
    createdAt?: Date;
    updatedAt?: Date;
}

export interface CreateServiceDto {
    title?: string;
    emoji?: string | undefined;
    description?: string | undefined;
    price?: number;
    isActive?: boolean;
    durationMinutes?: number;
}

export interface UpdateServiceDto {
    title?: string;
    emoji?: string | undefined;
    description?: string | undefined;
    price?: number;
    isActive?: boolean;
    durationMinutes?: number;
}